Bluetooth: Updated PBAP Filters logic to avoid loss of fields.
In PBAP there are filters for unsupported fields. But the logic
to check these has issue. It checks for Attribute string holds
unsupported parameter, rather than begins with.  So if the
field data has pattern matching unsupported field, that complete
attribute is being removed. The current logic fixes the issue.

CRs-Fixed: 332412

Change-Id: I29beae6c083298540c0bf88edbc523f9a6274bdb