Bluetooth: Truncate the received OPP file name instead of rejecting it
Including extn of the file, Linux supports 255 character as a maximum length of the
file name to be created. Hence, Instead of sending OBEX_HTTP_INTERNAL_ERROR
as a response, truncate the length of the file name and save it. This check majorly
helps in the case of vcard, where Phone book app supports contact name to be saved
more than 255 characters, But the server rejects the card just because the length of
vcf file name received exceeds 255 Characters.

Change-Id: I65935e2dfc33800636455e18bd9e29bdba5d87f5
CRs-fixed: 334935