Bluetooth: OPP: Update the user confirmation with proper value
User confirmation for the Incoming Files is updated in the Array
from the Database. There are two Database reads, the first read
is used for evaluating the user confirmation and the second read
to update the information in the Local Array. The stored
information in the Array will be used to evaluate a change in
its value in the database. Because of this, if there is a change
in the database from the point of first read to second read, the
value stored in the array would be different than the value
evaluated. In this case the evaluation during next updateShare
call will not see any change in user confirmation value in database
and array and hence will not result in Notifying the ServerSession
which is waiting for user confirmation.

CRs-fixed: 346867

Change-Id: I8cd0f70ac91d0c6cbbc045d81ad6a7f13918a1ce