Fixed Proximity and Thermometer Client profile concurrency issue
Fixed the profiles to only create Bluetooth Gatt service for
the services they are interested in when they receive the
ACTION GATT intent. Otherwise, when proximity is launched first
and thermometer next, the proximity starts to create thermometer
profile service on receiving the ACTION GATT (Broadcast intent) and
the callback on discover thermometer services goes to the proximity app.

Change-Id: I40a7cfc93567bbefff1803c1878b3e87acb2c7a0
CRs-fixed: 354108