Bluetooth: Do not terminate the session when a Batch contains an empty file
If an empty file is rejected by the remote device, client session is
terminated and remaining files in the batch also marked as failure.
This change continues to send the pending files in the batch even an
empty file is rejected by the remote device. Also it provides appropriate
message to the user.

Change-Id: I0eeef0b430909bd3041edfb0d32add21ec9a4a4d
CRs-fixed: 353630