Allow an empty file information to be stored in ContentValues
OPP implementation doesn't store an empty file information in
ContentValues though the file has a valid URI info and it is allowed
to transfer an empty file to the remote device. Due to which, in OPP
file transfer history it became an unknown file. The change enables
OPP to update the valid empty file info in ContentValues which can be
processed by a ContentResolver.

Change-Id: Ie853f094270f566e09c27ca7a19923d402369396
CRs-fixed: 403471