Bluetooth: Added new Handler to delay incomming Sco connection
Found that when there is active call and try to connect to s11 headset it
disconnects the SCO, does roleSwitch and then sends connectSco before we
finish handling the disconnectSco request this is reason DUT always rejects
the incoming connectSco request. Now with this fix made sure connectSco is
delayed and handled after the disconnectSco.

CRs-fixed: 423824
Change-Id: If994020636e714f8842dc5e191c0173449684fd0