Bluetooth: Moved the HFP state to proper state when is turned off
Found issue in some cases when Bluetooth is Turned Off the HFP
profile is still in Connecting. Because of this when the BT turned
On again the the Device shows as CONNECTING in Settings Menu. Now
this change makes sure HFP profile is moved to DISCONNECTED in this
case.

Change-Id: Iebfafe3ff0c1202fec53c6ed4a83897c99de646a
CRs-Fixed: 452187